Day 6

+++ src/common/mod.rs
@@ -1,12 +1,12 @@
 use itertools::Itertools;
 
-pub fn read_integers(input: &str) -> Option<Vec<i32>> {
+pub fn read_integers(input: &str) -> Option<Vec<i128>> {
     input.split_whitespace()
         .map(|line| line.parse().ok())
         .collect()
 }
 
-pub fn read_labeled_integers(input: &str) -> Option<Vec<(&str, i32)>> {
+pub fn read_labeled_integers(input: &str) -> Option<Vec<(&str, i128)>> {
     input.split_whitespace()
         .tuples::<(&str, &str)>()
         .map(|(a, b)| match b.parse() {
