+++ src/day5/mod.rs
@@ -4,7 +4,7 @@ use std::iter;
 
 use itertools::Itertools;
 
-pub fn solver(star: u8) -> fn(String) -> i32 {
+pub fn solver(star: u8) -> fn(String) -> i128 {
     match star {
         1 => star1,
         2 => star2,
@@ -12,16 +12,16 @@ pub fn solver(star: u8) -> fn(String) -> i32 {
     }
 }
 
-fn star1(input: String) -> i32 {
+fn star1(input: String) -> i128 {
     let lines: Vec<_> = input.split_whitespace()
         .tuples::<(&str, &str, &str)>()
         .map(|(a, _, b)| (v(a), v(b)))
         .collect();
 
-    let mut map: HashMap<(i32, i32), i32> = HashMap::new();
+    let mut map: HashMap<(i128, i128), i128> = HashMap::new();
 
     for ((x1, y1), (x2, y2)) in lines {
-        let l = (cmp::max(i32::abs(x2 - x1), i32::abs(y2 - y1)) + 1) as usize;
+        let l = (cmp::max(i128::abs(x2 - x1), i128::abs(y2 - y1)) + 1) as usize;
         let positions: Vec<_> = if x1 == x2 {
             let x = iter::repeat(x1).take(l);
             let y = num::range_step_inclusive(y1, y2, num::signum(y2 - y1));
@@ -41,27 +41,27 @@ fn star1(input: String) -> i32 {
 
     map.values()
         .filter(|&&v| v >= 2)
-        .count() as i32
+        .count() as i128
 }
 
-fn v(input: &str) -> (i32, i32) {
+fn v(input: &str) -> (i128, i128) {
     input.split(",")
-        .map(|s| s.parse::<i32>().unwrap())
-        .tuples::<(i32, i32)>()
+        .map(|s| s.parse::<i128>().unwrap())
+        .tuples::<(i128, i128)>()
         .next()
         .unwrap()
 }
 
-fn star2(input: String) -> i32 {
+fn star2(input: String) -> i128 {
     let lines: Vec<_> = input.split_whitespace()
         .tuples::<(&str, &str, &str)>()
         .map(|(a, _, b)| (v(a), v(b)))
         .collect();
 
-    let mut map: HashMap<(i32, i32), i32> = HashMap::new();
+    let mut map: HashMap<(i128, i128), i128> = HashMap::new();
 
     for ((x1, y1), (x2, y2)) in lines {
-        let l = (cmp::max(i32::abs(x2 - x1), i32::abs(y2 - y1)) + 1) as usize;
+        let l = (cmp::max(i128::abs(x2 - x1), i128::abs(y2 - y1)) + 1) as usize;
         let positions: Vec<_> = if x1 == x2 {
             let x = iter::repeat(x1).take(l);
             let y = num::range_step_inclusive(y1, y2, num::signum(y2 - y1));
@@ -83,5 +83,5 @@ fn star2(input: String) -> i32 {
 
     map.values()
         .filter(|&&v| v >= 2)
-        .count() as i32
+        .count() as i128
 }
