+++ src/template/mod.rs
@@ -1,6 +1,6 @@
 use crate::common;
 
-pub fn solver(star: u8) -> fn(String) -> i32 {
+pub fn solver(star: u8) -> fn(String) -> i128 {
     match star {
         1 => star1,
         2 => star2,
@@ -8,10 +8,10 @@ pub fn solver(star: u8) -> fn(String) -> i32 {
     }
 }
 
-fn star1(input: String) -> i32 {
+fn star1(input: String) -> i128 {
     0
 }
 
-fn star2(input: String) -> i32 {
+fn star2(input: String) -> i128 {
     0
 }
